Clean up and fuck up

+++ src/snake.h
@@ -78,7 +78,7 @@ typedef struct Geometry
   char density_filepath[LINE_LEN];
   char opacity_table_filepath[LINE_LEN];
   int icycle;
-  size_t nz_cells;
+  int nz_cells;
   double tot_tau;
   double T_init;
   double T_disk;
